company_researcher/requirements.txt
phidata>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.2
python-multipart>=0.0.6
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
openai>=1.6.1
pandas>=2.1.4
numpy>=1.24.0
beautifulsoup4>=4.9.1
lxml>=4.2.1
ollama
streamlit>=1.29.0